parcel_agent/requirements.txt
crewai>=0.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
requests>=2.28.0